have the frontend for cli, it is able to detect my voice but havent gotten jarvis to respond, the frontend can only do what jarvis tools, need to add orchestrator to have more tools
