+++ .jarvis_active_sessions.json
@@ -1,17 +1,1 @@
-{
-  "budget": {
-    "connected": true
-  },
-  "jarvis": {
-    "connected": true
-  },
-  "search": {
-    "connected": true
-  },
-  "system": {
-    "connected": true
-  },
-  "trading": {
-    "connected": true
-  }
-}
+{}
