+++ .jarvis_history
@@ -668,3 +668,12 @@
 
 # 2025-09-25 18:18:55.163802
 +trading.get_balance
+
+# 2025-09-26 11:28:58.877433
++start-api
+
+# 2025-09-26 12:04:00.166616
++list
+
+# 2025-09-26 12:08:33.322345
++start-api
